change the fb303_core namespace back to facebook.fb303
Summary:
While eventually it would be nice to change this namespace to just `fb303`,
using just `fb303` will require doing some significant codemods to update
fb303 users.  For now it's easiest to keep the namespace as `facebook.fb303`
to let us move forward with open sourcing fb303.  We can come back and
codemod the namespace to `fb303` in a subsequent diff after moving some more
of the code into this file.

The main issue with using `fb303` is that it creates ambiguity with the
existing `facebook.fb303` namespace still used elsewhere.  Defining a
top-level `fb303` namespace will require many parts of the code be updated to
explicitly resolve this ambiguity.

Reviewed By: wez

Differential Revision: D13828563

fbshipit-source-id: 8252de497288ebe2cb377f1fd3a4c0e81cdc298b