break ServiceData dependency on common/time
Summary:
To open source ServiceData and its dependencies, the dependency on common/time's WallClockUtil::NowInSecFast must be broken. Fortunately, NowInSecFast is equivalent to `std::time`, so introduce a new function that wraps `std::time` (and the faster CLOCK_GETTIME on macOS).

Eventually, we plan to make fb303 and the stats code consistently use std::chrono, but that's a bigger change across all of fbcode.

Reviewed By: yfeldblum, simpkins

Differential Revision: D15226138

fbshipit-source-id: c1bfa0ea827cec2bd3afcc11cd6e023facc15965