Make getCounters() skip queuing in Thrift server and run in a separate thread pool.
Summary:
Currently when `getCounters()` is sent to Thrift server, it can be blocked by an unhealthy thread pool, causing observibility issues when the server is overloaded, and the service owner wants to check out the counters to see what happened.
To make `getCounters()` unaffected, firstly it's annotated as an event base request, so that it skips waiting in in ThreadManagers, then a separated executor is added to run such requests.

Reviewed By: yfeldblum

Differential Revision: D17172755

fbshipit-source-id: e06c6ad3f9818246a6b6850e554add75410be3f2