build and install a python library from the fb303 thrift interface
Summary:
Update CMakeLists.txt to also generate python sources from the thrift
interface, and install it as a library that can be used by
`add_fb_python_executable()` in downstream projects.

The steps required to install both the C++ and python components of the
library are slightly complicated.  In the future I would like to add an
`install_fbthrift_library()` helper function that takes care of this more
easily.

Reviewed By: chadaustin

Differential Revision: D17127614

fbshipit-source-id: 6905f52ccec77107177cf5afccb3f87d85b8ed92