Make getRegexCounters() run in separate thread pool.
Summary: Currently when a `getRegexCounters()` request is sent to Thrift server, it can be blocked by a stucked thread pool, which used to be the case for `getCounters()` as well. We can unblock the requests in a similar way to `getCounters`, which is assigning the requests to a separate thread pool than the normal ones.

Reviewed By: yfeldblum

Differential Revision: D17818684

fbshipit-source-id: f7cdec5de422797fe55d043d135911db0b8eaf92