fb303: ThreadLocalStats: allow passing durations without export types
Summary:
This adds a new constructor for `ThreadLocalStats` that allows passing
durations without passing export types (currently, we have a constructor that
allows passing neither, and one that requires both).

This is used later in this stack for  `common/rust/stats`, which calls
`exportStat` on `ThreadLocalStats` instances, instead of passing export types
through the constructor.

Reviewed By: chadaustin

Differential Revision: D18268516

fbshipit-source-id: 80eeecb9e186e706cca4e02394489d0532b7013b