Make ThreadCachedServiceData a LeakySingleton
Summary:
ThreadCachedServiceData seems to be designed as 'folly::Singleton' but we have a lot of code across the codebase that does not check whether this Singleton is available or not, so lots of `ThreadCacheServiceData::get()->XXX` calls.
To accomodate these callsites we can just make this thing a `folly::LeakySingleton`. and only convert the underlying publisher thread to `folly::Singleton`.

Reviewed By: andriigrynenko

Differential Revision: D24693897

fbshipit-source-id: d6c67ca5bc4f53f3e798b40dfc88b660becf7919