Use shared locks and atomic variables in TLTimeseriesT
Summary:
## Memory

Replacing `folly::MicroLock` (1 byte) with `folly::SharedMutex` (4 bytes) does not change the size of the `TLTimeseriesT` class. It remains 96 bytes just as before this diff because of memory alignment.

## CPU
As stated in `folly/MicroLock.h` when space is not important usage of MicroLock is discouraged in favour of `std::mutex`.

`folly::SharedMutex` has similar performance with the `std::` version if there is not write contention albeit much smaller size - 4 vs 56 bytes.

Therefore in the case when the `TLTimeseriesT` is used as intended (from a single thread) this diff should produce no CPU regressions:
- the locks are slightly cheaper
- but the atomic variables are slightly more expensive

## Why this diff?

Frequently `TLTimeseriesT` cannot be used as intended. For example in libraries the number of threads that are going to call a certain method is not known in advance. If one writes a thread-safe method the user might call them from 1 or 1000 threads.

Hence for instrumenting the library one cannot create a  `TLTimeseriesT` per thread easily.

To avoid the bottleneck of an exclusive lock we propose using atomic variables and shared locks:
- `aggregate` is called rarely (every 1 sec or more) hence most of the time the lock will not be exclusively held
- the threads that call `addValue` can do it in parallel

Reviewed By: yfeldblum

Differential Revision: D24689901

fbshipit-source-id: 09775684bfbe3bf745c1e575394ed61d8e2f5b4f