Remove seperate logic for first value in histogram
Summary:
Currently there is logic that keeps histograms which only have one value reported to it separately.

The multi level histogram keeps the histogram for several time buckets. Just in case we had only one value added, it will keep reporting the value even if it was reported long ago and shouldn't appear in the time window level we are querying.

This is inconsistent with non histogram metrics (`avg` for example) reported from the same histogram, which don't have this logic.

A scenario would look like this:
1. Report value 5 to histogram
2. Immediately query the 60 second level -> `avg` reports 5, `p50` reports 5.
3. Wait 2 minutes
4. look at the 60 second level -> `avg` reports 0 (no value in the last 60 seconds) but `p50` keeps reporting 5.

I think this does more harm than good, so I suggest to kill this behavior.

Reviewed By: simpkins

Differential Revision: D26339016

fbshipit-source-id: 120485f8b78876e60ce66a534edcf86e7d9a836d