Move to F14NodeMap
Summary:
getThriftFuncHistParams() returns a pointer to an element in the map,
which is technically unsafe for F14FastMap. While the current usage of that
returned pointer is likely safe today, that could change over time. Instead,
move to F14NodeMap which provides stability of pointers to elements in the map.

Reviewed By: mogeb

Differential Revision: D27948698

fbshipit-source-id: 51944849d2bae7d8a48f073c08381adaf3ad895c