getdeps optionally can get hg info from env var
Summary:
Adds an environment variable to getdeps to provide `hg` info to avoid calling `hg` directly.

When using `getdeps` inside a containerized environment (which we need to build Research Super Cluster tooling with the correct linker attributes), `getdeps` fails because of unregistered mercurial extensions in the `hgrc`.

This allows `getdeps` to be useable in an environment where the mercurial extensions used in a project aren't installed/available.

Differential Revision: D34732506

fbshipit-source-id: 2b0fc664cb4afe1315ff62d35c617ae1e9a126e3