use libcurl in getdeps
Summary:
When using getdeps inside of a container, Python's urllib isn't able to download from dewey lfs (see this post for details https://fb.workplace.com/groups/systemd.and.friends/permalink/2747692278870647/).

This allows for getdeps to use `libcurl` to fetch dependencies, which allows for a getdeps build to work inside the container environment.

Differential Revision: D34696330

fbshipit-source-id: 5bdd78bb25a63540de6bda380ca3beb472c34c4d