Allow clients to limit number of fb303 counters read.
Summary:
Some clients are not in a position to consume all the counters that a server can potentially send
in reponse to getCounters.  This change introduces a thrift header, fb303_counters_limit, that clients can
set to constrain the number of counters that are returned from the server. Counters over this limit are dropped
by the server and not sent to the client.

Reviewed By: sruffell

Differential Revision: D35760589

fbshipit-source-id: 4f0e58446022e89e167da7b9d58a8cb7afcc502a