Allow clients to limit the number of counters read using getRegexCounters
Summary:
This is a follow up of D35760589 (https://github.com/facebook/fb303/commit/99cb8603a7d02e65a80d4f81515363ae67abdb3d). Some clients are not in a position to consume all the counters that a server can potentially send
in reponse to getRegexCounters. Even though this is a rare case, some regular expressions like ".*" can send all the counters available.  This change introduces a thrift header, fb303_counters_limit, that clients can
set to constrain the number of counters that are returned from the server during getRegexCounters call. Counters over this limit are dropped
by the server and not sent to the client.

Reviewed By: sruffell

Differential Revision: D36790108

fbshipit-source-id: 059f27b164ded426d667687856aa06fac0cceaca