Adding dirty keys flag to fb303 counter maps - preparatory diff for regex caching optimization change
Summary: This is a preparatory diff for D39002862. It converts all string keyed map used by callback counters, dynamic counters and service data counters to use a struct that also contains a dirty bit flag. This dirty bit will be used by diff D39002862 to signal whether a map has new or modified counters (which will be used to determine whether cache needs to be refreshed)

Differential Revision: D35058781

fbshipit-source-id: 380a785661dce4d7180b638cd1a0e4df09371e4c