- Use dirty bit to track when new keys are added or deleted to counters/ dynamic counters or callback map
Summary: This is part of a stack of diffs to optimize getRegexCounters with caching on the server side. A dirty bit is added to keep track of when a new counter is added to fb303 maps. This will be used by getRegexCounters to determine whether or not the cache of matching keys is valid.

Differential Revision: D39593116

fbshipit-source-id: 719bd6015c6b5086069f8b4cbba6e1b8395ccb7a