fb3030: Enable regex caching optimization via Thrift headers
Summary: fb303 regex caching is a major change and must be deployed gradually.  This change lets us to turn the optimization on a per call basis by adding an entry to the Thrift headers for getRegexCounters().  This will allow us to run canary tests by selectively enabling the optimization on certain hosts for a second fb303 request that can be compared with the first, by results returned.

Reviewed By: yfeldblum

Differential Revision: D39003576

fbshipit-source-id: fbf87b99e5a47f635b2cd42bd07ed8fb92e2589f