Pin all git deps to a rev
Summary:
All `branch` and `tag`s have been replaced with an exact `rev` to avoid
the risk of having one of those symbolic refs change from under us.
Also, these existing cases set a bad example that others tend to copy -
so remove the temptation.

This was done manually by going to each repo and seeing what the branch
or tag resolved to.

Reviewed By: dtolnay

Differential Revision: D43329883

fbshipit-source-id: c437a3174d660f4aed7d3dcbfb63d3cd0ce433da