Give ThreadLocal in ThreadCacheServiceData custom tags.
Summary:
The tag determines the underlying StaticMeta class that is used to
manage lifecyle of instances of that type. Defaulting to void can cause
contention in create/destroy of different counter types and with other
unrelated thread locals. With separate tags, any contention is at least easier
to investigate as it can only come from its own type.

Custom tag for TLStats types is the main one we need. I added it for the other
less used thread locals too and if we want to minimize use of custom tags, then
we can keep it just for the TL stat objects.

Reviewed By: sarangbh

Differential Revision: D43512986

fbshipit-source-id: bc342e66a0b8245105241e669b96117741fcaf27