Revert D43512986: Multisect successfully blamed D43512986 for test or build failures
Summary:
This diff is reverting D43512986
D43512986: Give ThreadLocal in ThreadCacheServiceData custom tags. by gargn has been identified to be causing the following test or build failures:

Tests affected:
- [hphp/facebook/test:verify_facebook_interp_jit - hphp/facebook/test/ext/ext_fci.php](https://www.internalfb.com/intern/test/281475060707007/)

Here's the Multisect link:
https://www.internalfb.com/intern/testinfra/multisect/1595139
Here are the tasks that are relevant to this breakage:
T114752638: 4 tests started failing for oncall hphp_hphpi in the last 2 weeks
We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

Differential Revision: D43541676

fbshipit-source-id: c132daed2cdb020641eec8b782676a7e3f8df1d5