let DynamicQuantileStatMap use FormattedKeyHolder and ThreadLocal
Summary:
Similar to like `DynamicTimeseriesWrapper` and `DynamicHistogramWrapper`.

The current algorithm with hazard-pointers is susceptible to O(N^2) time-complexity ramp-up. We can continue to accelerate the mapping from the subkeys to the stat objects by accelerating key-formatting with a thread-local cache in `FormattedKeyHolder` and then by accelerating the map from formatted-key to stat with another thread-local cache.

Reviewed By: sethdelliott

Differential Revision: D47070971

fbshipit-source-id: 8b6cd0691b90025b09a4b5c5e973490101158e8d