fix size types in MultiLevelTimeseries ctor
Summary: The size types are `int` (unexpected) but they are forwarded to the base-class ctor which takes `size_t` (expected). Downstream callers pass `size_t` (expected) but run into violations of warninng `shorten-64-to-32`. This violation was papered over with an explicit cast, but it is better to fix the ctor size types.

Reviewed By: chadaustin, meleshuk

Differential Revision: D47472384

fbshipit-source-id: e6ef383b8005b81a77b25587a6b4aac6936dc32a