Update libbpf-sys snapshot
Summary:
The sched_ext team uses libbpf-rs for their atropos user space scheduler. They
would like to make use of features of libbpf for which no release has been cut
yet. The libbpf maintainer would like to wait a bit more until before creating a
new release. The version of libbpf that libbpf-rs relies on is controlled by
whatever is vendored with libbpf-sys. Without a new release of libbpf, there
will not be a new release of libbpf-sys with a new libbpf snapshot either.
Hence, to still get the desired functionality from libbpf, with this change we
pull in a patched version of libbpf-sys that vendors libbpf at commit
5a46421ad837e876197295844696884c8587852a (from 2023-08-24). This is a temporary
and we will switch back to relying on official releases once with libbpf and
libbpf-sys have released versions that include the desired functionality.

Reviewed By: Decave

Differential Revision: D48698179

fbshipit-source-id: 2f1d0887f55e5f44e684bc165061a7bb4fe4adaa