Relax requirements for thread-local stats
Summary:
Thread-local stats currently require that they are used explicitly by a specific thread. This is actually much stronger than needed, all we need is serialization. The current requirements prevent the use of these stats on user-level thread abstractions that can migrate between OS threads, for example `EventBase`s.

This diff relaxes the contract. There is no logic change in non-debug builds: the only logic change is in the debug assertions, which are adjusted accordingly. Even if we didn't have these assertions, nowadays TSAN would reliably detect incorrect use.

As an added benefit, this removes the need for the `swapThreads()` method, which based on the comments next to the invocations I'm removing is quite confusing.

Reviewed By: a-square

Differential Revision: D53127894

fbshipit-source-id: c9c867028d5973ea27ff690852faa3671f211155