Delete unused services_crate option and {{program:servers}}
Summary:
Prior to {D54500949}, the generated `:foo-rust` target used to depend on `:foo-rust-services` and needed to refer to its contents.

After that diff, there is no longer any generated code that depends on the `-services` crate, so we no longer need to pass the Thrift compiler a crate name by which to refer to the contents of the `-services` crate. This diff cleans that up.

Reviewed By: shayne-fletcher

Differential Revision: D56726365

fbshipit-source-id: d0e1cc1d302a6e848699c0e2e78692c9c6f9af5f