Add a dirty bit to TL TimeSeriesType
Summary:
If no calls to `addValue()` happen between two `reset()` calls, the second `reset()` is unnecessarily doing a few expensive atomic RMWs (including acquiring and releasing a mutex) when there is in fact nothing to reset.

We can avoid this by having `addValue()` maintain a dirty bit. This requires slightly changing the algorithm, because the side bit and dirty bit must be handled atomically, or `reset()` can end up clearing the dirty bit even though some new writes have arrived in the meantime.

To enable this we pack the writer "lock" bit, the side bit, and the dirty bit into a single state word. `reset()` is allowed to modify the state only during a quiescent state (and acquire/release synchronization is sufficient to guarantee sequential consistency here), so it is safe to unset the dirty bit and switch sides.

Reviewed By: bmaurer

Differential Revision: D57437818

fbshipit-source-id: 0fe0357285f214b872dbecc44947b4452fa119aa