autocargo: fix ambiguous type
Summary:
the thrift library target `fbcode//thrift/annotation:all` defines `thrift_srcs = []`. in this case, a 'thrift_build.rs' will be generated that contains the line `let srcs = &[];`. this is of ambiguous type. fix this by ascription: `let srcs: &[&str]`.

before:
```counterexample
$ buck2 run fbcode//common/rust/cargo_from_buck/autocargo:autocargo && (cd fbcode/thrift/lib/rust/annotation/all && THRIFT1=$(buck2 build fbcode//thrift/compiler:thrift --show-full-simple-output) CARGO_TARGET_DIR=~/tmp cargo build)

error[E0282]: type annotations needed for `&[_; 0]`
  --> thrift_build.rs:41:9
   |
41 |     let srcs = &[];
   |         ^^^^   --- type must be known at this point
```

the `autocargo-verification` job should be expected to fail here. this is because it's `autocargo` itself that is being updated by this diff.

Reviewed By: Imxset21

Differential Revision: D57521067

fbshipit-source-id: 02f697a63898d5a27a611938d185ff4642edab2f