let FormattedKeyHolder cache the local last item
Summary: It may be the case that it is sufficiently likely for the last element found to be the next element looked up. If so, having it cached can remove the cost of map-lookup, including indirections and other chunk-walking calculations. If not, checking the cache would be of minimal cost.

Reviewed By: DenisYaroshevskiy

Differential Revision: D59171277

fbshipit-source-id: ac700c8b434b5a869c70f02fd99752d1951fcf04