perf: cut indirection in TimeseriesWrapper::tcTimeseries
Summary:
The fast-path access-path is via `ThreadLocal<shared_ptr<TLTimeseries>>`. This looks up the pointer to the `shared_ptr` in the internal `ThreadLocalPtr` and then has to indirect through that `shared_ptr`.

We can remove the indirection by changing the fast-path access-path to `ThreadLocalPtr<TLTimeseries>`. This lookups up the pointer to the `TLTimeseries` in the `ThreadLocalPtr`. The trick, then, is holding onto a copy of the `shared_ptr` in the deleter.

Reviewed By: skrueger

Differential Revision: D59342239

fbshipit-source-id: 0c8d33de41c839aa96c257a4418d9000ff13cb91