Create an enum for superconsole interaction toggles
Summary:
Current implementation of console toggles relies on `char` and display is error-prone (.e.g we forgot to add `x` option to the Help). I initially tried to fix this in D60960924 but then realized that this could happen in the future. Better to make it type-safe so that we get it for free (via `strum::EnumIter`).

Also rename  `ConsoleInteraction` to `SuperConsoleInteraction`  to signify that these only for for superconsole.

Reviewed By: stepancheg

Differential Revision: D60965981

fbshipit-source-id: e0008b46dd752446b45877e6f0c2acf43f528c7b