Use callback to initialize the histogram (#56)
Summary:
Pull Request resolved: https://github.com/facebook/fb303/pull/56

Removing expensive unused temporary.

Internal

TLHistogramT<LockTraits>::initGlobalStat is 787k :
https://fburl.com/strobelight/zj0f8fqw
{F1803643308}

Almost oll of this is HistogramBuckets
https://fburl.com/strobelight/a0mwtylr
{F1803646026}

If we look into distribtuions, it's mostly construction/destruction - no copies.
https://fburl.com/scuba/strobelight_services/t2yhalrb
 {F1803647007}

So - hypothesis is that most of the time the temporary is constructed but unused.

In fact, we can also confirm it by looking at `initGlobalStat` profile - all of the time is in constructing and destroying the temporary, not copying.
https://fburl.com/scuba/strobelight_services/g2iurhvz
 {F1803649266}

So - changing the api to make construction lazy.

Differential Revision: D61207391