let FormattedKeyHolder cache the local last item
Summary: It may be the case that it is sufficiently likely for the last element found to be the next element looked up. If so, having it cached can remove the cost of map-lookup, including indirections and other chunk-walking calculations. If not, checking the cache would be of minimal cost.

Reviewed By: ddrcoder

Differential Revision: D61810053

fbshipit-source-id: 9ff03aa6a5bf1a46fac1b72fd784985bbb52eb04