enable user input to prevent request context being persisted
Summary:
Currently, TLStatsAsyncAggregator::scheduleAggregation will call AsyncTimeout::scheduleTimeout which saves as a default parameter the current RequestContext. The possible issue with this is that RequestContext can possibly persist indefinitely until program end since on the timer expiration, the class calls scheduleTimeout, this cycle will continue.

This issue has happened in other areas too, such as HHWheelTimer. Simplest solution is to allow callers to provide input to scheduleAggregation if desired to have control over the RequestContext behavior. This change is made s.t. the existing behavior should not be affected.

Differential Revision: D65911229