prevent TLStatsAsyncAggregator from saving RequestContext to AsyncTimeout
Summary:
Currently, TLStatsAsyncAggregator::scheduleAggregation will call AsyncTimeout::scheduleTimeout which saves as a default parameter the current RequestContext. The possible issue with this is that RequestContext can possibly persist indefinitely until program end since on the timer expiration, the class calls scheduleTimeout, this cycle will continue.

This issue has happened in other areas too, such as HHWheelTimer. Since this class's counters operate on a global scope, it does not make sense to carry any request specific information into the timeout. Thus, for the invocations of scheduleTimeout, we pass in nullptr for the RequestContext.

The test case now shows that the RequestContext object is no longer persisted into the timeout.

Update: introduced member function scheduleAggregationTimeout to handle the calling of scheduleTimeout. Name helps avoid name clashing with AsyncTimeout and is a bit more descriptive.

Reviewed By: ot

Differential Revision: D65911229

fbshipit-source-id: 552493f7957425a9a0cfeb2661a53d619871999f