cut the option to skip the regex-match-cache
Summary: The regex-match-cache (`folly::RegexMatchCache`), used in `ServiceData` and in associated classes, is stable. It is mandatory in the bookkeeping path and it is on-by-default in the query path for all services. But there is no longer any need to have the option to avoid it in the query path.

Reviewed By: dmm-fb

Differential Revision: D64484083

fbshipit-source-id: 67dd1886f1a6a0d015d3ac4a82590ba34fbe1650