revise ThreadCachedServiceData per-key export caching
Summary:
Cut `ThreadCachedServiceData::ExportCacheMap` etc. It's an extra map that every call to `ThreadCachedServiceData::addStatValue` has to traverse, and the map lookup to find the export status can be just as costly as the map lookup to find the stat.

Instead, just cache the export status in the thread-local stat-map value.

Some aspects of the `ExportType` overloads of `addStatValue` and `clearStat` have seemingly strange behavior. But, for now, we attempt to preserve the current behavior.

Differential Revision: D68659181

fbshipit-source-id: f9028696f9007d0e2bafeddd35551f1663ff6149