Use QuantileStats for function counters
Summary:
Currently, there are counters like `time_process_us` that offer `SUM` / `AVG`. Explicitly, services can also opt in to track p50, p90, p99 percentiles via `exportThriftFuncHist()` (backed by `folly::Histogram`).

**This diff enables a new set of `QuantileStat` backed counters, on by default, that provide various percentiles for all thrift methods.**

The plan is to eventually replace `folly::Histogram` with `QuantileStat`. `folly::Histogram` has bad memory performance, which is presumably why it's only opt-in via `exportThriftFuncHist()`. `QuantileStat` also comes with greater accuracy, but has a small CPU cost.

Note that `TServerCounters` already makes use of `QuantileStat`.

For backwards compatibility, we're keeping the current `exportThriftFuncHist()` mechanism. Once this rolls in:
* all the quantiles that are not manually exported via the above function call will be collected
* we'll switch the opted-in counters from `folly::Histogram` to `QuantileStat`. This maintains the same counter names, and should be invisible to the services.
* additional percentiles will be available for the previously manually exported histograms (p10, p95)

We're using the singular `TStatsPerThread::TimeSeries::addValue()` point to either add to the `folly::Histogram` or the new `QuantileStat`.

Note that a key change is that there is no need to aggregate per-thread values for `QuantileStat`. We can add the values normally from each thread context, and internally this will aggregate on the first call.

With the added bonus that they provide sum / avg, it entirely removes the need for `TimeSeries` objects for the quantiles that will be replaced, so those will be tracked by `QuantileStat`-s directly.

Guarding everything behind a killswitch which is defaulted to `true`.

Reviewed By: ot

Differential Revision: D72855344

fbshipit-source-id: ef03f6c979953b14e426af67b7d672e357b6dc04