Do not pull quantile stat out of TStatsPerThread::TimeSeries
Summary:
This addresses the TODO.

The intent was to pull the `QuantileStat` object up one level; semantically it makes more sense for this to be owned by `TStatsPerThread`.

However, the way the callers interact with this is by calling `addValue()`, as they are oblivious to quantile stats being involved or only the sum / count increased:
https://www.internalfb.com/code/fbsource/fbcode/common/fb303/cpp/TFunctionStatHandler.cpp?lines=73-74

https://www.internalfb.com/code/fbsource/[4d5449951f9e]/fbcode/fb303/TFunctionStatHandler.cpp?lines=434-435

etc.

With a change of ownership, the callers would have to call an `addValue()`-equivalent on `TStatsPerThread` which is easy to miss and erorr prone. So encapsulating the object under this interface is likely preferable.

Reviewed By: ot

Differential Revision: D75142806

fbshipit-source-id: a39883fc9597f2e8f0c484f17c910bd3565ed7f3