Bump python-setuptools source version
Summary: We want to make use of this for our OSS build of CinderX (see diffs further up the stack). However, the older version of setuptools seems to break in really weird ways I can't find work arounds for. Upgrading it seems to make most of the problems go away or avoidable so if nobody complains I'd rather fix things by just doing this.

Reviewed By: alexmalyshev

Differential Revision: D79195099

fbshipit-source-id: 2d65e8960e435536dc7dd7e6a142f1bd2df89a41