Allow users to specifcy `now`
Summary:
We are changing the underlying fb303 timeseries to have sub-second buckets. It requires a sub-second clock to reap the benefits. However, there is a small performance regression if we change `now()` when calling `addStatValue` for all ServiceData. This is because of this optimization in the MultilevelTimeseries https://fburl.com/code/904ox41z where it caches sum and count until the clock moves to the next tick. A more granular clock makes this optimization less useful.

In this change, we allow users of ServiceData to specify `now` which defaults to a clock of second granularity to make this change perf neutral.

#buildall

Reviewed By: islamismailov

Differential Revision: D82854336

fbshipit-source-id: 2c3d8ab2c0e91ac4c96ecbd54d7a574a6b6c7872