Make TLStat registration lock optimized
Summary:
This diff optimizes the stat registration in ThreadLocalStats by using a lock-free data structure to store it as pending stats.
The pending stats are store in linked list and new stats gets added to head of the list atomically.
This list gets drained into container on unlink and aggregate calls.

Reviewed By: yfeldblum

Differential Revision: D84789193

fbshipit-source-id: cd06a7369057d458b1424c9fa28028aaf2ab5103