RegexUtils re-encapsulations
Summary:
Re-encapsulates that the counter-map and the regex-match-cache are always modified together, and symmetrically, with `cachedAddString` and `cachedEraseString`. Simplifes all call-sites.

Imposes a requirement where the counter-map is actualy a set where the string keys are embedded in the set values, that the map's key-equal function also serve double-duty as a function for getting references to the string keys. Does not impose any requirement when the counter-map is a map type. This requirement enables restoring the encapsulation and simplifying all the call-sites without all the call-sites having explicitly to pass around get-key-ref functions as arguments to `cachedAddString` and `cachedEraseString`.

Additional:
* Switch to CallbackValuesMap to F14VectorSet, a vector-set, which stores all of its elements in a single array (aka vector). This type is preferable when iteration is dominant over lookup, which is the case here.
* Avoid allocations in `CallbackValuesMap::getValues`, almost always.

Reviewed By: ot

Differential Revision: D84833685

fbshipit-source-id: 6da7da1f9f8002443984ee21bfac71a138153ddc