const-qualify DynamicOption{G,S}etter
Summary:
These callbacks can be called concurrently, so it's better to require that at least their state is `const` (they can still mutate external state, but it's more self-evident that that needs to be synchronized). This is not a restrictive requirement, since lambdas are const by default, in fact no call site passed a non-const function.

While at it, we can remove the other `const_cast`s in the file by using the `Synchronized::asNonConstUnsafe()` which is intended precisely for this use case, without side-stepping const correctness.

Reviewed By: uvdn7

Differential Revision: D85246269

fbshipit-source-id: ffcee462080bdaf33fbc21d42a7ae34c9e6b60f9