Adding per storage/shard/view quantiles for delta reads
Summary:
# Context
https://docs.google.com/document/d/1AyDcFtJKRT49dmxNigpjm8-0zS0EaHZn4722yc2Qfuc

S564141 follow up. We need better visibility into our storage interactions.

# Problem
We have added a lot of per storage quantiles (per shard) but are missing information at the view-level. It would be helpful for debugging if we had view-level shard quantiles.

# Solution/this diff
Adding quantiles for view-level delta read bytes/rows. This required injecting a `ViewProvider` into `Persister` in order to build this map of view prefix to stats. We could technically build this map as we iterate over the delta records that are returned, but this would require a lock.

One thing to note is that for every delta we will now have to iterate through the key (using `KeyDeserializer`) to get the view prefix. I don't expect this to be expensive, but I'll check for any impact to persister latency in alphas.

Note that the quantiles are P50, P90, P99, and P100. SUM and AVG are intentionally missing because these counters are very expensive for ads where we have thousands of views. When I originally tested this in ads_alpha with 8 different counter types, we exceeded the limit for the number of counters we could export from the host.

Differential Revision: D86000285

fbshipit-source-id: 0c49765a478d953c764743292b1135ddc271cd42