Make getdeps Git fetcher use branch in directory name to avoid clashes
Summary: If we have two manifests which checkout from the same repository but at different branches, one of them will break when reusing the same checkout in the getdeps cache. This is because we only do a shallow checkout but default and so only the branch originally used to populatd the cache is available.

Reviewed By: alexmalyshev

Differential Revision: D87041221

fbshipit-source-id: f54cdbc1392c791d9da3ad3bb3236c354e2a0837