Fix OSS builds (removes 3.12mp build)
Summary:
This diff was generated by UTC AI agent for your convenience on top of T.

**This diff was generated by UTC AI agent on behalf of jbower for T. If the diff quality is poor, consider contacting the user to provide clearer instructions on the task.**

- If you are happy with the changes, commandeer it if minor edits are needed. (**we encourage commandeer to get the diff credit**)
- If you are not happy with the changes, please comment on the diff with clear actions and send it back to the author. Racer will pick it up and re-generate.
- If you really feel the UTC AI Agent is not helping with this change (alas, some complex changes are hard for AI) feel free to abandon this diff.
I think it's reasonable to not have an OSS build of 3.12 - in practice nobody is ever going to use this oddity as it requires Meta Python. CinderX with stock CPython is what we expect going forwards.

I'm trying to reconcile a few constraints here without making the code too complicated:
* For PGO with LLVM we need to have llvm-ar installed, but this isn't available on Sandcastle or Ubuntu Latest (on GitHub).
* It's easyish to install GCC on Sandcastle.
* GCC-13 is available on Ubuntu Latest in GitHub but installing another version is annoying because our GitHub workflow for getdeps is autogenerated.
* GCC-13 is sufficient for CinderX + 3.14 but GCC-14 is needed for CinderX + 3.12 (something seems to have been fixed around headers for atomics when mixing C + C++).

So, making GCC-13 a minimum requirement for OSS builds is good enough to get PGO builds working on Sandcastle + GitHub provided we only target 3.14+.

Reviewed By: itamaro

Differential Revision: D88570106

fbshipit-source-id: 7151be4bda9f754d296bf56f20fba90615918a54